Split widget: effects panel
